solving file i/o

+++ 0x15-file_io/2-append_text_to_file.c
@@ -7,6 +7,7 @@
  *
  * Return: On success, returns 1. On failure, returns -1.
  */
+
 int append_text_to_file(const char *filename, char *text_content)
 {
 	int fd, _write;
@@ -21,15 +22,16 @@ int append_text_to_file(const char *filename, char *text_content)
 	if (!text_content)
 	{
 		close(fd);
-		return (1);
+		return(1);
 	}
+
 	_write = write(fd, text_content, strlen(text_content));
-	if (_write == -1 || _write != strlen(text_content))
+	if (_write == -1 || _write != (int)strlen(text_content))
 	{
 		close(fd);
 		return (-1);
 	}
 
 	close(fd);
-	return (1);
+	return(1);
 }
